+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Stripe SDK Changelog
 
+## 1.0.0
+
+Improve API for SCA-related features
+Improve examples
+
+## 0.0.2
+
+* Remove typed models
+* Add some support for payment intents
+* Complete support for payment methods
+* Complete support for tokens
+* Add examples
+* Major cleanup
+
 ## 0.0.1+2
 
 * Add analysis_options.yaml